pyproject.toml
[tool.poetry]
name = "onnxmodeldeployment"
version = "0.1.0"
description = ""
authors = ["Gabriel Leite Bessa <gabrielleitebessa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"
onnxruntime = "^1.18.0"
numpy = "1.26.4"
pillow = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
